Additional updates to grand park PHP site

+++ php/actions/clearcity.php
@@ -1,18 +1,26 @@
 <?php
 include("actionbase.php");
 
-$event = $_GET["city"];
 
-$tripID = GetSingleValueDB("SELECT TripId FROM Trip WHERE UserId = " . $userid . " AND Event = '" . $event . "';","TripId");
+if($userid == 0){
+    //not a logged in user
+    $_SESSION["eventTimes"] = [];
+    $_SESSION["itineraryTimes"] = [];
+}else {
+    $event = $_GET["event"];
 
-$eventSQL = "DELETE FROM TripEvent Where TripId = " . $tripID . ";";
-ExecuteSQL($eventSQL);
-$itinerarySQL = "DELETE FROM TripItinerary Where TripId = " . $tripID . ";";
-ExecuteSQL($itinerarySQL);
+    $tripID = GetSingleValueDB("SELECT TripId FROM Trip WHERE UserId = " . $userid . " AND Event = '" . $event . "';","TripId");
 
+    $eventSQL = "DELETE FROM TripEvent Where TripId = " . $tripID . ";";
+    ExecuteSQL($eventSQL);
+    $itinerarySQL = "DELETE FROM TripItinerary Where TripId = " . $tripID . ";";
+    ExecuteSQL($itinerarySQL);
+}
 
 if (isset($_SERVER["HTTP_REFERER"])) {
     echo "<script>location='" .$_SERVER["HTTP_REFERER"] . "'</script>"; /* Redirect browser */
+}else{
+    echo "<script>location='index.php'</script>"; /* Redirect browser */
 }
 
 ?>
